+++ test/fixtures/sample/lib/sample.dart
@@ -1,4 +1,5 @@
 import 'package:collection/collection.dart';
+import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:path/path.dart' as p;
 
 import 'src/internal_for_typedef.dart';
@@ -9,6 +10,10 @@ export 'src/exported.dart';
 typedef Handler = void Function(TypedefThing thing);
 
 Handler? handler;
+
+String get _viaGetter => used();
+
+String get exposedViaGetter => _viaGetter;
 
 String runSample() {
   final items = [used(), p.basename('a/b')];